fix analytics merge issue
