+++ lib/presentation/settings/sections/analytics_section.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ribn/constants/strings.dart';
-import 'package:ribn/presentation/analytics_provider.dart';
+import 'package:ribn/providers/analytics_provider.dart';
 import 'package:ribn_toolkit/constants/styles.dart';
 import 'package:ribn_toolkit/widgets/atoms/custom_toggle.dart';
